+++ src/App.css
@@ -16,10 +16,6 @@
   margin-left: auto;
   margin-right: auto;
   max-width: 800px;
-  min-height: 500px;
-  background-color: #111;
-  display: flex;
-  justify-content: center;
   align-items: center;
 }
 
@@ -100,4 +96,8 @@ form button:hover a {
   max-width: 800px;
   margin: auto;
   margin-top: 5px;
+}
+
+.chakra-form-control {
+  margin-bottom: 8px;
 }
